MEN-3513: Test that R/O rootfs checksum is correct also at runtime.
Move some of the runtime helpers from test_update.py into a helpers.py
file, so we can use it from other tests.

Changelog: None
